+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/settings/UserSettings.kt
@@ -49,7 +49,9 @@ class UserSettingsViewImpl : UserSettingsView {
                     AppearanceSettingsButton(userSettingsViewModel)
                     PrivacySettingsButton(userSettingsViewModel)
                     DeviceSettingsButton(userSettingsViewModel)
-                    ProfilesSettingsButton(userSettingsViewModel)
+                    if (userSettingsViewModel.canShowProfilesSettings) {
+                        ProfilesSettingsButton(userSettingsViewModel)
+                    }
                 }
             }
         }