+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/timeline/TimelineViewModel.kt
@@ -302,7 +302,7 @@ class TimelineViewModelImpl(
                 .filter(timelineElementViewModelFactorySelector::supports)
                 .take(100)
                 .scan(0) { count, _ -> count + 1 }
-                .debounce(100)
+                .debounce(100.milliseconds)
                 .map {
                     when {
                         it in 1..99 -> it.toString()
@@ -411,6 +411,7 @@ class TimelineViewModelImpl(
     }
 
     init {
+        outerScope.removeMarkAsUnread()
         updateReadEventAndMarker()
 
         initTimeline()
@@ -1060,13 +1061,20 @@ class TimelineViewModelImpl(
                     .onFailure { log.error(it) { "cannot set read marker for event $nextReadUntil in $nextReadUntilRoomId" } }
                     .onSuccess { log.debug { "successfully set read marker for message: $nextReadUntil in $nextReadUntilRoomId" } }
             }
-            launch {
+            removeMarkAsUnread()
+
+        }.join()
+    }
+
+    private fun CoroutineScope.removeMarkAsUnread() {
+        this.launch {
+            if (matrixClient.room.getAccountData(roomId, MarkedUnreadEventContent::class).map { it?.unread }
+                    .firstOrNull() ?: false) {
                 matrixClient.api.room.setAccountData(MarkedUnreadEventContent(false), roomId, userId)
                     .onFailure { log.warn(it) { "could not reset unread in $roomId" } }
                     .onSuccess { log.debug { "successfully reset unread in $roomId" } }
             }
-
-        }.join()
+        }
     }
 
     private val getReceiptsByEventCache = concurrentMutableMap<RoomId, Flow<Map<EventId, Set<UserId>>>>()